Fix in rollup stats - add missing conversion to double between division.

+++ h2o-core/src/main/java/water/fvec/RollupStats.java
@@ -185,8 +185,9 @@ class RollupStats extends Iced {
     double delta = _mean - rs._mean;
     if (_rows == 0) { _mean = rs._mean;  _sigma = rs._sigma; }
     else {
-      double a = _rows/(_rows + rs._rows);
-      double b = rs._rows/(_rows + rs._rows);
+      // still converting long to double, most likely ok, can be avoided...
+      double a = (double)_rows/(_rows + rs._rows);
+      double b = (double)rs._rows/(_rows + rs._rows);
       _mean = (_mean*a + rs._mean*b);
       _sigma = _sigma + rs._sigma + delta*delta * _rows*rs._rows / (_rows+rs._rows);
     }
